+++ tang-tool/tang-generator/src/main/java/com/tang/generator/utils/TableUtils.java
@@ -1,9 +1,13 @@
 package com.tang.generator.utils;
 
 import com.google.common.base.CaseFormat;
+import com.tang.commons.utils.Assert;
 import com.tang.commons.utils.SpringUtils;
 import com.tang.generator.autoconfigure.GeneratorProperties;
 import com.tang.generator.entity.GenTable;
+import org.springframework.lang.NonNull;
+
+import java.util.Objects;
 
 /**
  * 表工具类
@@ -22,7 +26,9 @@ public class TableUtils {
      *
      * @param table 表信息
      */
-    public static void initTable(GenTable table) {
+    public static void initTable(@NonNull GenTable table) {
+        Assert.isNull(table, "表信息不能为空");
+        Assert.isNull(GENERATOR_PROPERTIES.getPackageName(), "包路径不能为空");
         table.setClassName(getClassName(table.getTableName()));
         table.setPackageName(GENERATOR_PROPERTIES.getPackageName());
         table.setModuleName(getModuleName(GENERATOR_PROPERTIES.getPackageName()));
@@ -47,7 +53,8 @@ public class TableUtils {
      * @param packageName 包路径
      * @return 模块名
      */
-    private static String getModuleName(String packageName) {
+    private static String getModuleName(@NonNull String packageName) {
+        Assert.isNull(packageName, "包路径不能为空");
         var packageNames = packageName.split("\\.");
         return packageNames[packageNames.length - 1];
     }
@@ -60,6 +67,9 @@ public class TableUtils {
      */
     private static String getClassName(String tableName) {
         var className = tableName;
+        if (Objects.isNull(GENERATOR_PROPERTIES.getRemovePre())) {
+            return CaseFormat.LOWER_UNDERSCORE.to(CaseFormat.UPPER_CAMEL, className);
+        }
         var prefixes = GENERATOR_PROPERTIES.getRemovePre().split(",");
         for (String prefix : prefixes) {
             if (className.startsWith(prefix)) {